* move quickPick helpers to utils * remove dashboard launch (will do in separate PR)
